+++ src/main/java/delfos/io/csv/dataset/item/DefaultContentDatasetToCSV.java
@@ -197,8 +197,8 @@ public class DefaultContentDatasetToCSV implements ContentDatasetToCSV {
 
             int j = 40499;
             while (reader.readRecord()) {
-                Float latitude;
-                Float longitude;
+                Double latitude;
+                Double longitude;
 
                 int idItem;
 
@@ -210,13 +210,13 @@ public class DefaultContentDatasetToCSV implements ContentDatasetToCSV {
                 }
 
                 if (latitudeColumn != null) {
-                    latitude = Float.parseFloat(reader.get(latitudeColumn));
+                    latitude = Double.parseDouble(reader.get(latitudeColumn));
                 } else {
                     latitude = null;
                 }
 
                 if (longitudeColumn != null) {
-                    longitude = Float.parseFloat(reader.get(longitudeColumn));
+                    longitude = Double.parseDouble(reader.get(longitudeColumn));
                 } else {
                     longitude = null;
                 }
